fix(zentao): fix test errors
